+++ src/XRoadLib/Serialization/Mapping/IntegerTypeMap.cs
@@ -15,7 +15,7 @@ namespace XRoadLib.Serialization.Mapping
         public override object Deserialize(XmlReader reader, IXmlTemplateNode templateNode, ContentDefinition content, XRoadMessage message)
         {
             if (reader.IsEmptyElement)
-                return MoveNextAndReturn(reader, HandleEmptyElement(content, message));
+                return reader.MoveNextAndReturn(HandleEmptyElement(content, message));
 
             var value = reader.ReadElementContentAsString();
             if (string.IsNullOrEmpty(value))