+++ src/XRoadLib/Serialization/Mapping/OptimizedContentTypeMap.cs
@@ -1,5 +1,7 @@
 ﻿using System.IO;
 using System.Xml;
+using System.Xml.Linq;
+using XRoadLib.Extensions;
 using XRoadLib.Schema;
 using XRoadLib.Serialization.Template;
 
@@ -16,7 +18,21 @@ namespace XRoadLib.Serialization.Mapping
 
         public object Deserialize(XmlReader reader, IXmlTemplateNode templateNode, ContentDefinition content, XRoadMessage message)
         {
-            if (!reader.ReadToDescendant("Include", NamespaceConstants.XOP))
+            if (reader.IsEmptyElement)
+                return DeserializeBase64Content(reader, message);
+
+            if (!reader.ReadToContent())
+            {
+                if (reader.NodeType == XmlNodeType.EndElement)
+                    return GetEmptyAttachmentStream(message);
+
+                throw new InvalidQueryException("Invalid content element.");
+            }
+
+            if (reader.NodeType != XmlNodeType.Element)
+                return DeserializeBase64Content(reader, message);
+
+            if (!reader.MoveToElement(reader.Depth, XName.Get("Include", NamespaceConstants.XOP)))
                 throw new InvalidQueryException("Missing `xop:Include` reference to multipart content.");
 
             var contentID = reader.GetAttribute("href");
@@ -28,6 +44,31 @@ namespace XRoadLib.Serialization.Mapping
                 throw new InvalidQueryException($"MIME multipart message does not contain message part with ID `{contentID}`.");
 
             return attachment.ContentStream;
+        }
+
+        private static Stream GetEmptyAttachmentStream(IAttachmentManager attachmentManager)
+        {
+            var tempAttachment = new XRoadAttachment(new MemoryStream()) { IsMultipartContent = false };
+            attachmentManager.AllAttachments.Add(tempAttachment);
+            return tempAttachment.ContentStream;
+        }
+
+        private static object DeserializeBase64Content(XmlReader reader, XRoadMessage message)
+        {
+            if (reader.IsEmptyElement)
+                return reader.MoveNextAndReturn(GetEmptyAttachmentStream(message));
+
+            const int bufferSize = 1000;
+
+            int bytesRead;
+            var buffer = new byte[bufferSize];
+
+            var contentStream = GetEmptyAttachmentStream(message);
+
+            while ((bytesRead = reader.ReadContentAsBase64(buffer, 0, bufferSize)) > 0)
+                contentStream.Write(buffer, 0, bytesRead);
+
+            return contentStream;
         }
 
         public void Serialize(XmlWriter writer, IXmlTemplateNode templateNode, object value, ContentDefinition content, XRoadMessage message)