+++ src/XRoadLib/Serialization/Mapping/TypeMap.cs
@@ -16,11 +16,5 @@ namespace XRoadLib.Serialization.Mapping
         public abstract object Deserialize(XmlReader reader, IXmlTemplateNode templateNode, ContentDefinition content, XRoadMessage message);
 
         public abstract void Serialize(XmlWriter writer, IXmlTemplateNode templateNode, object value, ContentDefinition content, XRoadMessage message);
-
-        protected static object MoveNextAndReturn(XmlReader reader, object value)
-        {
-            reader.Read();
-            return value;
-        }
     }
 }