+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+#### 1.3.8 - Unreleased
+
+* Fix: duplicate complex type definitions in WSDL document.
+* Fix: deserialize MTOM/XOP enabled property as inline base64 content when multipart attachment content id is missing.
+
 #### 1.3.7 - 21.12.2018
 
 * Check namespaces of complex type elements.